IB/hfi1: Unregister mmu callbacks from the correct mm
Jira: TOSS-3119

The hfi1 driver registers a mmu_notifier callback when /dev/hfi1 is opened,
and unregisters it when the device is closed.  The driver assumes that the
close will always happen from the same context as the open, which is
incorrect.  In particular, closes due to SIGKILL or OOM killer activity may
happen from a different context.  In these cases, the wrong mm is passed to
mmu_notifier_unregister(), which causes improper reference counting for the
victim mm, and eventual memory corruption.

The mmu_rb_handler structure is modified to cache the mm used during
registration, so that that mm can be passed at unregistration time.

Signed-off-by: Jim Foraker <foraker1@llnl.gov>