WIP: Make sure SDMA resources are freed
The previous patch to clean up the locking resulted in sdma_rb_remove being
called with the SDMA_CACHE_NODE_EVICT bit set.  Initially the mm pointer was
being used to differentiate if sdma_rb_remove was being called by the mmu
notifier or not.  However, by cleaning up the semantics to always pass an mm
object this code broke.

For now this just adds a flag which indicates if we are being called by the mmu
notifier or not.  If we are then the check for race with the evict code is
valid.

It seems like the we should be able to remove this bit check but I will have to
inspect the code more.  For now I believe this restores the original logic.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>